# shader_jit_x64_compiler: Remove ABI overhead of LG2 and EX2