--- a/src/video_core/shader/shader_jit_x64_compiler.h
+++ b/src/video_core/shader/shader_jit_x64_compiler.h
@@ -106,6 +106,13 @@ private:
      */
     void FindReturnOffsets();
 
+    /**
+     * Emits data and code for utility functions.
+     */
+    void CompilePrelude();
+    Xbyak::Label CompilePrelude_Log2();
+    Xbyak::Label CompilePrelude_Exp2();
+
     const std::array<u32, MAX_PROGRAM_CODE_LENGTH>* program_code = nullptr;
     const std::array<u32, MAX_SWIZZLE_DATA_LENGTH>* swizzle_data = nullptr;
 
@@ -120,6 +127,9 @@ private:
 
     using CompiledShader = void(const void* setup, void* state, const u8* start_addr);
     CompiledShader* program = nullptr;
+
+    Xbyak::Label log2_subroutine;
+    Xbyak::Label exp2_subroutine;
 };
 
 } // Shader