We reimplement log2f and exp2f in assembly.

Algorithms were inspired by those found in Mesa I've used a higher order approximation for exp2 than they did for more accuracy.

Performance numbers (for 1.2 billion sequential calls of the same instruction on my machine):

| Instruction | Old time | New time |
| ------------- | ------------- |  ------------- |
| `ex2` | 21.5s | 8.8s |
| `lg2` | 13.3s | 7.3s |

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/citra-emu/citra/3145)
<!-- Reviewable:end -->