@yuriks 2017-11-26
Nice perf wins and nice, tests!

---

Reviewed 2 of 2 files at r1, 2 of 2 files at r2, 1 of 1 files at r3.
Review status: all files reviewed at latest revision, 9 unresolved discussions.

---

*[src/tests/CMakeLists.txt, line 24 at r1](https://reviewable.io:443/reviews/citra-emu/citra/3145#-Kzp2ZnPZg5mXpWlkcvZ:-Kzp2ZnQP-e20YGIa1AN:buerzos) ([raw file](https://github.com/citra-emu/citra/blob/e2b22c721a353e8b954fa5d123ee215f9afb8a55/src/tests/CMakeLists.txt#L24)):*
> ```CMake
> target_link_libraries(tests PRIVATE ${PLATFORM_LIBRARIES} catch-single-include nihstro-headers Threads::Threads)
> if (ARCHITECTURE_x86_64)
>     target_link_libraries(tests PRIVATE xbyak)
> ```

That's weird, I would've expected this to be pulled transitively via video_core, so you shouldn't need to specify it here, since you're not using any xbyak functions directly.

Ah, looks like we're specifying xbyak as a PRIVATE dependency in common and video_core's CMakeLists. I think this is wrong, since they're used in public headers from those libraries, so they should be changed to PUBLIC.

---

*[src/tests/video_core/shader/shader_jit_x64_compiler.cpp, line 5 at r1](https://reviewable.io:443/reviews/citra-emu/citra/3145#-Kzp3tLV8mdB3tf5GeFW:-Kzp3tLW9aU9-TWzUFdS:bnq1e51) ([raw file](https://github.com/citra-emu/citra/blob/e2b22c721a353e8b954fa5d123ee215f9afb8a55/src/tests/video_core/shader/shader_jit_x64_compiler.cpp#L5)):*
> ```cpp
> // Refer to the license.txt file included.
> 
> #ifdef ARCHITECTURE_x86_64
> ```

Can you conditionally include the entire via on CMakeLists instead (like is done in the main codebase). Or do you have a particular reason to do it this way?

---

*[src/tests/video_core/shader/shader_jit_x64_compiler.cpp, line 49 at r1](https://reviewable.io:443/reviews/citra-emu/citra/3145#-Kzp5OCYSwkYdQVJw5Bf:-Kzp5OCYSwkYdQVJw5Bg:bw1hl0h) ([raw file](https://github.com/citra-emu/citra/blob/e2b22c721a353e8b954fa5d123ee215f9afb8a55/src/tests/video_core/shader/shader_jit_x64_compiler.cpp#L49)):*
> ```cpp
>     });
> 
>     const auto run = [&](float input) {
> ```

You could encapsulate this (compiling the shader, setting up N inputs and the output, and then the `run` function receiving N params and returning the output) into a standalone class to avoid duplicating this setup twice, and since it seems to be widely applicable to other shader tests.

---

*[src/video_core/shader/shader_jit_x64_compiler.cpp, line 930 at r2](https://reviewable.io:443/reviews/citra-emu/citra/3145#-KzplIGKEQoGqujW-X-R:-KzplIGL7lsWBTY5_sPC:b-43bd4e) ([raw file](https://github.com/citra-emu/citra/blob/d9383f7c78a5d9d3f5a634a78369a1c4a74a783e/src/video_core/shader/shader_jit_x64_compiler.cpp#L930)):*
> ```cpp
> void JitShader::CompilePrelude() {
>     CompilePrelude_Log2();
>     CompilePrelude_Exp2();
> ```

Optional: Have these two functions return the Labels to the compiled function, which are then assigned here, making it a bit more obvious that's where their results are going.

---

*[src/video_core/shader/shader_jit_x64_compiler.cpp, line 969 at r2](https://reviewable.io:443/reviews/citra-emu/citra/3145#-Kzp8OGF2U5S3UnBvOE7:-Kzp8OGF2U5S3UnBvOE8:bl96ih2) ([raw file](https://github.com/citra-emu/citra/blob/d9383f7c78a5d9d3f5a634a78369a1c4a74a783e/src/video_core/shader/shader_jit_x64_compiler.cpp#L969)):*
> ```cpp
> 
>     align(16);
>     L(input_is_nan);
> ```

Any reason not to merge this block with one of the other two ret from the other two cases?

---

*[src/video_core/shader/shader_jit_x64_compiler.cpp, line 980 at r2](https://reviewable.io:443/reviews/citra-emu/citra/3145#-KzphS9H9lZQf1XVCnnX:-KzphS9IHFrpJ1MHZOdO:bnih6mq) ([raw file](https://github.com/citra-emu/citra/blob/d9383f7c78a5d9d3f5a634a78369a1c4a74a783e/src/video_core/shader/shader_jit_x64_compiler.cpp#L980)):*
> ```cpp
> 
>     align(16);
>     log2_subroutine = getCurr();
> ```

You should be able to also use labels for the routine entry points and then use call with them, letting xbyak possibly emit a shorter relative call instead.

---

*[src/video_core/shader/shader_jit_x64_compiler.cpp, line 993 at r2](https://reviewable.io:443/reviews/citra-emu/citra/3145#-Kzp8tBeZn9pFRr5tmXb:-Kzp8tBfTDU3g-Z5Kgq3:btsc1vp) ([raw file](https://github.com/citra-emu/citra/blob/d9383f7c78a5d9d3f5a634a78369a1c4a74a783e/src/video_core/shader/shader_jit_x64_compiler.cpp#L993)):*
> ```cpp
>     and_(eax, 0x7f800000);
>     and_(edx, 0x007fffff);
>     movss(SCRATCH, xword[rip + c0]); // Preload c0.
> ```

Curious: How did you determine placement of this load? (i.e. why here, instead of at the beginning of this block?)

---

*[src/video_core/shader/shader_jit_x64_compiler.cpp, line 1050 at r2](https://reviewable.io:443/reviews/citra-emu/citra/3145#-Kzp8f3mTJY_aIqrHOAG:-Kzp8f3mTJY_aIqrHOAH:bnpxnua) ([raw file](https://github.com/citra-emu/citra/blob/d9383f7c78a5d9d3f5a634a78369a1c4a74a783e/src/video_core/shader/shader_jit_x64_compiler.cpp#L1050)):*
> ```cpp
> 
>     align(16);
>     L(ret_label);
> ```

Isn't it better to just put ret_label before the final ret in the subroutine, instead of dedicating 16 bytes for it?

---

*[src/video_core/shader/shader_jit_x64_compiler.cpp, line 783 at r3](https://reviewable.io:443/reviews/citra-emu/citra/3145#-KzqWNrmiavMQu2_D_fZ:-KzqWNrmiavMQu2_D_f_:b-ec4zgo) ([raw file](https://github.com/citra-emu/citra/blob/af6d0acd23145dc79aef002c8080263230c3ea13/src/video_core/shader/shader_jit_x64_compiler.cpp#L783)):*
> ```cpp
>     add(ABI_PARAM2, static_cast<Xbyak::uint32>(offsetof(UnitState, registers.output)));
>     CallFarFunction(*this, Emit);
>     if (Common::GetCPUCaps().avx) {
> ```

After some research, it seems that the latest versions of gcc, clang and MSVC all generate vzeroupper before calls or rets in functions in which AVX was used. Tests: https://godbolt.org/g/K7Eeun (Note that at the time of writing the MSVC in godbolt is out of date and has a bug.)

Given that, do we still need to emit this? And in any case, would it make more sense to gate this behind __AVX__ (the compiler define) instead of the runtime cap bit? Presumably if __AVX__ isn't set then we don't need to worry about AVX code in the rest of the application.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/citra-emu/citra/3145#-:-KzqXJM5Kglu3Riip7nu:bmvymp6)*
<!-- Sent from Reviewable.io -->